nonlinear: switch to Metropolis
Has less than half the run time as before, with not much effect on the
(dismal) effective sample sizes.
There is somewhat arbitrary constant, 1, for the minimum sd of the
metropolis jumps.  I was probably above the bound for this run.
Added acceptance rate to the last column of MCTRACE.
Appears to be getting reasonable numbers.