nonlinear: major fix to threshhold model
I forgot to update how the values were imputed and so the previous
commit's code was bogus.  Because the posterior z will always preserve
the value of y, the usual way of handling imputation breaks down for
the.  I modfied the code to do something that seemed sensible and might
be justifiable.
Effective sample sizes are quite a bit lower than previous (bogus)
run's; now most are around 75, with a couple parameters around 25.